# libs/@local/graph/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "graph",
    "description": "HASH Graph API",
    "contact": {
      "name": "HASH"
    },
    "license": {
      "name": "AGPL-3"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/accounts": {
      "post": {
        "tags": [
          "Graph",
          "Account"
        ],
        "operationId": "create_account",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertAccountIdParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The schema of the created account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UntaggedActorId"
                }
              }
            }
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
    "/data-types": {
      "post": {
        "tags": [
          "Graph",
          "DataType"
        ],
        "operationId": "create_data_type",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDataTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The metadata of the created data type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaybeListOfDataTypeMetadata"
                }
              }
            }
          },
          "409": {
            "description": "Unable to create data type in the store as the base data type URL already exists"
          },
          "422": {
            "description": "Provided request body is invalid"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      },
      "put": {
        "tags": [
          "Graph",
          "DataType"
        ],
        "operationId": "update_data_type",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDataTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The metadata of the updated data type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataTypeMetadata"
                }
              }
            }
          },
          "404": {
            "description": "Base data type ID was not found"
          },
          "422": {
            "description": "Provided request body is invalid"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
...
    "/entities": {
      "post": {
        "tags": [
          "Graph",
          "Entity"
        ],
        "operationId": "create_entity",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The created entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            }
          },
          "404": {
            "description": "Entity Type URL was not found"
          },
          "422": {
            "description": "Provided request body is invalid"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      },
      "patch": {
        "tags": [
          "Graph",
          "Entity"
        ],
        "operationId": "patch_entity",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEntityParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            }
          },
          "404": {
            "description": "Entity ID or Entity Type URL was not found"
          },
          "422": {
            "description": "Provided request body is invalid"
          },
          "423": {
            "description": "The entity that should be updated was unexpectedly updated at the same time"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
...
    "/entity-types": {
      "post": {
        "tags": [
          "Graph",
          "EntityType"
        ],
        "operationId": "create_entity_type",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntityTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The metadata of the created entity type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaybeListOfEntityTypeMetadata"
                }
              }
            }
          },
          "400": {
            "description": "Provided request body is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "409": {
            "description": "Unable to create entity type in the datastore as the base entity type ID already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "500": {
            "description": "Store error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Graph",
          "EntityType"
        ],
        "operationId": "update_entity_type",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The metadata of the updated entity type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityTypeMetadata"
                }
              }
            }
          },
          "404": {
            "description": "Base entity type ID was not found"
          },
          "422": {
            "description": "Provided request body is invalid"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
...
    "/property-types": {
      "post": {
        "tags": [
          "Graph",
          "PropertyType"
        ],
        "operationId": "create_property_type",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The metadata of the created property type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaybeListOfPropertyTypeMetadata"
                }
              }
            }
          },
          "409": {
            "description": "Unable to create property type in the store as the base property type ID already exists"
          },
          "422": {
            "description": "Provided request body is invalid"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      },
      "put": {
        "tags": [
          "Graph",
          "PropertyType"
        ],
        "operationId": "update_property_type",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The metadata of the updated property type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyTypeMetadata"
                }
              }
            }
          },
          "404": {
            "description": "Base property type ID was not found"
          },
          "422": {
            "description": "Provided request body is invalid"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDataTypeRequest": {
        "type": "object",
        "required": [
          "schema",
          "ownedById",
          "relationships",
          "provenance",
          "conversions"
        ],
        "properties": {
          "conversions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Conversions"
            }
          },
          "ownedById": {
            "$ref": "#/components/schemas/OwnedById"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvidedOntologyEditionProvenance"
          },
          "relationships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataTypeRelationAndSubject"
            }
          },
          "schema": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DataType"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DataType"
                }
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "CreateEntityRequest": {
        "type": "object",
        "required": [
          "ownedById",
          "entityTypeIds",
          "properties",
          "draft",
          "relationships",
          "provenance"
        ],
        "properties": {
          "confidence": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              }
            ]
          },
          "decisionTime": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Timestamp"
              }
            ]
          },
          "draft": {
            "type": "boolean"
          },
          "entityTypeIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VersionedUrl"
            }
          },
          "entityUuid": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EntityUuid"
              }
            ]
          },
          "linkData": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LinkData"
              }
            ]
          },
          "ownedById": {
            "$ref": "#/components/schemas/OwnedById"
          },
          "properties": {
            "$ref": "#/components/schemas/PropertyObjectWithMetadata"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvidedEntityEditionProvenance"
          },
          "relationships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityRelationAndSubject"
            }
          }
        },
        "additionalProperties": false
      },
      "CreateEntityTypeRequest": {
        "type": "object",
        "required": [
          "schema",
          "ownedById",
          "relationships",
          "provenance"
        ],
        "properties": {
          "ownedById": {
            "$ref": "#/components/schemas/OwnedById"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvidedOntologyEditionProvenance"
          },
          "relationships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityTypeRelationAndSubject"
            }
          },
          "schema": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/EntityType"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityType"
                }
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "DataTypeMetadata": {
        "oneOf": [
          {
            "type": "object",
            "title": "OwnedDataTypeMetadata",
            "required": [
              "recordId",
              "ownedById",
              "temporalVersioning",
              "provenance"
            ],
            "properties": {
              "conversions": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/Conversions"
                }
              },
              "ownedById": {
                "$ref": "#/components/schemas/OwnedById"
              },
              "provenance": {
                "$ref": "#/components/schemas/OntologyProvenance"
              },
              "recordId": {
                "$ref": "#/components/schemas/OntologyTypeRecordId"
              },
              "temporalVersioning": {
                "$ref": "#/components/schemas/OntologyTemporalMetadata"
              }
            }
          },
          {
            "type": "object",
            "title": "ExternalDataTypeMetadata",
            "required": [
              "recordId",
              "fetchedAt",
              "temporalVersioning",
              "provenance"
            ],
            "properties": {
              "conversions": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/Conversions"
                }
              },
              "fetchedAt": {
                "$ref": "#/components/schemas/Timestamp"
              },
              "provenance": {
                "$ref": "#/components/schemas/OntologyProvenance"
              },
              "recordId": {
                "$ref": "#/components/schemas/OntologyTypeRecordId"
              },
              "temporalVersioning": {
                "$ref": "#/components/schemas/OntologyTemporalMetadata"
              }
            }
          }
        ]
      },
...
      "Entity": {
        "type": "object",
        "description": "A record of an entity that has been persisted in the datastore, with its associated metadata.\n\nAn [`Entity`] represents a real-world object, concept, or thing within the knowledge graph.\nIt contains structured data in the form of properties, optional link data for establishing\nrelationships with other entities, and comprehensive metadata that describes the entity's\nprovenance, types, temporal information, and more.\n\nEach entity is an instance of one or more [`EntityType`]s defined in the ontology. The\nrelationship is similar to objects and classes in object-oriented programming:\n- [`EntityType`]s define the schema, structure, and constraints that entities must follow\n- [`Entity`] instances contain actual data conforming to those schemas\n\nAn entity:\n- Is identified by a unique [`EntityId`]\n- Has one or more [`VersionedUrl`]s in its `entity_type_ids` field linking to its types\n- Contains a set of properties structured according to the schemas defined in its types\n- May have links to other entities, establishing relationships in the knowledge graph\n- Includes comprehensive metadata for tracking provenance, versioning, and confidence\n\n[`EntityType`]: crate::ontology::entity_type::EntityType\n[`VersionedUrl`]: crate::ontology::VersionedUrl",
        "required": [
          "properties",
          "metadata"
        ],
        "properties": {
          "linkData": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LinkData"
              }
            ]
          },
          "metadata": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "properties": {
            "$ref": "#/components/schemas/PropertyObject"
          }
        }
      },
...
      "EntityId": {
        "type": "string"
      },
...
      "InsertAccountIdParams": {
        "type": "object",
        "properties": {
          "accountId": {
            "$ref": "#/components/schemas/UntaggedActorId"
          }
        },
        "additionalProperties": false
      },
...
      "MaybeListOfDataTypeMetadata": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/DataTypeMetadata"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataTypeMetadata"
            }
          }
        ]
      },
      "MaybeListOfEntityTypeMetadata": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/EntityTypeMetadata"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityTypeMetadata"
            }
          }
        ]
      },
...
      "PatchEntityParams": {
        "type": "object",
        "required": [
          "entityId",
          "provenance"
        ],
        "properties": {
          "archived": {
            "type": "boolean"
          },
          "confidence": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              }
            ]
          },
          "decisionTime": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Timestamp"
              }
            ]
          },
          "draft": {
            "type": "boolean"
          },
          "entityId": {
            "$ref": "#/components/schemas/EntityId"
          },
          "entityTypeIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VersionedUrl"
            }
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyPatchOperation"
            }
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvidedEntityEditionProvenance"
          }
        },
        "additionalProperties": false
      },
...
      "Status": {
        "$ref": "./models/status.json"
      },
...
      "UntaggedActorId": {
        "type": "string",
        "format": "uuid"
      },
...
      "UpdateDataTypeRequest": {
        "type": "object",
        "required": [
          "schema",
          "typeToUpdate",
          "relationships",
          "provenance",
          "conversions"
        ],
        "properties": {
          "conversions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Conversions"
            }
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvidedOntologyEditionProvenance"
          },
          "relationships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataTypeRelationAndSubject"
            }
          },
          "schema": {
            "$ref": "#/components/schemas/UpdateDataType"
          },
          "typeToUpdate": {
            "$ref": "#/components/schemas/VersionedUrl"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateEntityTypeRequest": {
        "type": "object",
        "required": [
          "schema",
          "typeToUpdate",
          "relationships",
          "provenance"
        ],
        "properties": {
          "provenance": {
            "$ref": "#/components/schemas/ProvidedOntologyEditionProvenance"
          },
          "relationships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityTypeRelationAndSubject"
            }
          },
          "schema": {
            "$ref": "#/components/schemas/UpdateEntityType"
          },
          "typeToUpdate": {
            "$ref": "#/components/schemas/VersionedUrl"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}